incomplete draft of splitting home fragment
