+++ app/src/main/res/layout/fragment_home.xml
@@ -1,63 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
+
+<ListView xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/item_list"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:orientation="vertical">
-
-    <LinearLayout
-        android:id="@+id/linearLayout"
-        android:layout_width="match_parent"
-        android:layout_height="78dp"
-        android:background="@color/brown"
-        android:paddingVertical="24dp"
-        android:paddingHorizontal="32dp"
-        android:orientation="horizontal">
-
-        <com.google.android.material.textview.MaterialTextView
-            android:layout_width="0dp"
-            android:layout_height="match_parent"
-            android:layout_gravity="center"
-            android:layout_weight="1"
-            android:gravity="center_vertical"
-            android:text="Inventory"
-            android:textColor="@color/white"
-            android:textSize="22sp"
-            android:textStyle="bold" />
-
-        <ImageView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:src="@drawable/logo"
-            app:tint="@color/white" />
-    </LinearLayout>
-
-    <LinearLayout
-        android:id="@+id/dropdown_buttons"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal"
-        android:gravity="end"
-        android:padding="4dp">
-
-        <com.google.android.material.button.MaterialButton
-            android:id="@+id/filter_dropdown_button"
-            style="?android:attr/borderlessButtonStyle"
-            android:layout_width="30dp"
-            android:layout_height="30dp"
-            android:padding="0dp"
-            android:layout_marginEnd="8dp"
-            app:icon="@drawable/baseline_tune_24"
-            app:iconSize="30dp"
-            app:iconTint="@color/black" />
-    </LinearLayout>
-
-
-    <ListView
-        android:id="@+id/item_list"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:divider="@android:color/transparent"
-        android:dividerHeight="5.0sp"
-        android:padding="8dp"></ListView>
-</LinearLayout>
+    android:divider="@android:color/transparent"
+    android:dividerHeight="5.0sp"
+    android:padding="8dp"
+    />
