+++ app/src/main/res/layout/fragment_home_base.xml
@@ -1,0 +1,69 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical">
+    <FrameLayout
+        android:id="@+id/base_state_options"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:padding="4dp">
+
+        <com.google.android.material.button.MaterialButton
+            android:id="@+id/select_items_button"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="@string/select"
+            android:textSize="20sp"
+            android:padding="0dp"
+            android:textStyle="bold" />
+
+        <LinearLayout
+            android:id="@+id/base_state_tools"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:orientation="horizontal"
+            android:layout_gravity="end">
+
+            <com.google.android.material.button.MaterialButton
+                android:id="@+id/filter_dropdown_button"
+                style="?android:attr/borderlessButtonStyle"
+                android:layout_width="48dp"
+                android:layout_height="48dp"
+                android:padding="2dp"
+                android:layout_marginEnd="8dp"
+                app:icon="@drawable/baseline_tune_24"
+                app:iconSize="30dp"
+                app:iconTint="@color/black"
+                android:contentDescription="@string/filter_dropdown_description"/>
+
+            <com.google.android.material.button.MaterialButton
+                android:id="@+id/sort_by_alpha_button"
+                style="?android:attr/borderlessButtonStyle"
+                android:layout_width="48dp"
+                android:layout_height="wrap_content"
+                android:padding="2dp"
+                app:icon="@drawable/baseline_sort_by_alpha_24"
+                app:iconSize="30dp"
+                app:iconTint="@color/black"
+                android:contentDescription="@string/sort_button_description"/>
+
+            <com.google.android.material.button.MaterialButton
+                android:id="@+id/sort_order_toggle"
+                style="?android:attr/borderlessButtonStyle"
+                android:layout_width="48dp"
+                android:layout_height="wrap_content"
+                android:padding="2dp"
+                app:icon="@drawable/swap_vert_24px"
+                app:iconSize="30dp"
+                app:iconTint="@color/black"
+                android:contentDescription="@string/toggle_description"/>
+        </LinearLayout>
+
+    </FrameLayout>
+
+    <include layout="@layout/fragment_home"/>
+
+</LinearLayout>
+
